+++ src/decode/loaders/LoaderBase.ts
@@ -3,9 +3,9 @@ import {
   HalfFloatType,
   LinearFilter,
   LinearMipMapLinearFilter,
+  LinearSRGBColorSpace,
   Loader,
   LoadingManager,
-  NoColorSpace,
   RGBAFormat,
   SRGBColorSpace,
   Texture,
@@ -34,7 +34,8 @@ const getImage = (blob: Blob) => {
 }
 
 export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFloatType, GainMapDecoderMaterial>, TUrl> {
-  private renderer: WebGLRenderer
+  private _renderer: WebGLRenderer
+  protected _internalLoadingManager: LoadingManager
   /**
    *
    * @param renderer
@@ -42,7 +43,8 @@ export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFl
    */
   constructor (renderer: WebGLRenderer, manager?: LoadingManager) {
     super(manager)
-    this.renderer = renderer
+    this._renderer = renderer
+    this._internalLoadingManager = new LoadingManager()
   }
 
   /**
@@ -68,9 +70,9 @@ export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFl
       16,
       16,
       HalfFloatType,
-      NoColorSpace,
+      LinearSRGBColorSpace,
       material,
-      this.renderer
+      this._renderer
     )
   }
 
@@ -121,7 +123,7 @@ export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFl
       RGBAFormat,
       UnsignedByteType,
       1,
-      NoColorSpace
+      LinearSRGBColorSpace
     )
 
     gainMap.flipY = needsFlip